+++ Wox.Core.Data/Storage/UserSelectedRecordStorage.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using Newtonsoft.Json;
-using Wox.Infrastructure.Storage;
 using Wox.Plugin;
 
-namespace Wox.Infrastructure.Storage
+namespace Wox.Core.Data.Storage
 {
     public class UserSelectedRecordStorage : BaseStorage<UserSelectedRecordStorage>
     {
