+++ Wox.Core.Data/Storage/UserSettings/ProgramSource.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Wox.Infrastructure.Storage.UserSettings
+namespace Wox.Core.Data.Storage.UserSettings
 {
     [Serializable]
     public class ProgramSource
