+++ Wox.Core.Data/Wox.Core.Data.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{4FD29318-A8AB-4D8F-AA47-60BC241B8DA3}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Wox.Infrastructure</RootNamespace>
-    <AssemblyName>Wox.Infrastructure</AssemblyName>
+    <RootNamespace>Wox.Core.Data</RootNamespace>
+    <AssemblyName>Wox.Core.Data</AssemblyName>
     <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <SolutionDir Condition="$(SolutionDir) == '' Or $(SolutionDir) == '*Undefined*'">..\</SolutionDir>
@@ -49,26 +49,14 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="PeHeaderReader.cs" />
     <Compile Include="Storage\UserSettings\FolderLink.cs" />
-    <Compile Include="Unidecoder.Characters.cs" />
-    <Compile Include="ChineseToPinYin.cs" />
-    <Compile Include="HttpRequest.cs" />
     <Compile Include="Storage\BaseStorage.cs" />
-    <Compile Include="FuzzyMatcher.cs" />
-    <Compile Include="GlobalHotkey.cs" />
-    <Compile Include="HotkeyModel.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Storage\UserSelectedRecordStorage.cs" />
     <Compile Include="Storage\UserSettings\UserSettingStorage.cs" />
     <Compile Include="Storage\UserSettings\PluginHotkey.cs" />
     <Compile Include="Storage\UserSettings\ProgramSource.cs" />
     <Compile Include="Storage\UserSettings\WebSearch.cs" />
-    <Compile Include="Unidecoder.cs" />
-    <Compile Include="WindowsShellRun.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Wox.Core\Wox.Core.csproj">
