+++ Wox.Core/SpecialKeyState.cs
@@ -1,0 +1,10 @@
+﻿namespace Wox.Core
+{
+    public class SpecialKeyState
+    {
+        public bool CtrlPressed { get; set; }
+        public bool ShiftPressed { get; set; }
+        public bool AltPressed { get; set; }
+        public bool WinPressed { get; set; }
+    }
+}
