+++ Wox.Core/packages.config
