+++ Wox.Plugin.SystemPlugins/Program/ProgramSources/UserStartMenuProgramSource.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using Wox.Infrastructure.Storage.UserSettings;
+using Wox.Core.Data.Storage.UserSettings;
 
 namespace Wox.Plugin.SystemPlugins.Program.ProgramSources
 {
