+++ Wox.Plugin/ISystemPlugin.cs
