+++ Wox.Plugin/Wox.Plugin.csproj
@@ -45,8 +45,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AllowedLanguage.cs" />
+    <Compile Include="ISystemPlugin.cs" />
     <Compile Include="Properties\Annotations.cs" />
-    <Compile Include="ArrayExtensions.cs" />
     <Compile Include="IPlugin.cs" />
     <Compile Include="ISettingProvider.cs" />
     <Compile Include="PluginPair.cs" />
@@ -58,6 +58,12 @@
     <Compile Include="Query.cs" />
     <Compile Include="Result.cs" />
     <Compile Include="ActionContext.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Wox.Core\Wox.Core.csproj">
+      <Project>{de47afa8-08c5-40f6-aa23-9aa988219214}</Project>
+      <Name>Wox.Core</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
