+++ Wox.Tests/Wox.Tests.csproj
@@ -55,9 +55,13 @@
       <Project>{097B4AC0-74E9-4C58-BCF8-C69746EC8271}</Project>
       <Name>Python.Runtime</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Wox.Infrastructure\Wox.Infrastructure.csproj">
+    <ProjectReference Include="..\Wox.Core\Wox.Core.csproj">
+      <Project>{DE47AFA8-08C5-40F6-AA23-9AA988219214}</Project>
+      <Name>Wox.Core</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Wox.Core.Data\Wox.Core.Data.csproj">
       <Project>{4FD29318-A8AB-4D8F-AA47-60BC241B8DA3}</Project>
-      <Name>Wox.Infrastructure</Name>
+      <Name>Wox.Core.Data</Name>
     </ProjectReference>
     <ProjectReference Include="..\Wox.Plugin\Wox.Plugin.csproj">
       <Project>{8451ECDD-2EA4-4966-BB0A-7BBC40138E80}</Project>
