+++ Wox.UAC/Wox.UAC.csproj
@@ -109,9 +109,9 @@
     <AppDesigner Include="Properties\" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Wox.Infrastructure\Wox.Infrastructure.csproj">
+    <ProjectReference Include="..\Wox.Core.Data\Wox.Core.Data.csproj">
       <Project>{4fd29318-a8ab-4d8f-aa47-60bc241b8da3}</Project>
-      <Name>Wox.Infrastructure</Name>
+      <Name>Wox.Core.Data</Name>
     </ProjectReference>
     <ProjectReference Include="..\Wox.Plugin\Wox.Plugin.csproj">
       <Project>{8451ecdd-2ea4-4966-bb0a-7bbc40138e80}</Project>
