+++ Wox/Wox.csproj
@@ -256,9 +256,9 @@
       <Project>{de47afa8-08c5-40f6-aa23-9aa988219214}</Project>
       <Name>Wox.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Wox.Infrastructure\Wox.Infrastructure.csproj">
+    <ProjectReference Include="..\Wox.Core.Data\Wox.Core.Data.csproj">
       <Project>{4fd29318-a8ab-4d8f-aa47-60bc241b8da3}</Project>
-      <Name>Wox.Infrastructure</Name>
+      <Name>Wox.Core.Data</Name>
     </ProjectReference>
     <ProjectReference Include="..\Wox.Plugin.SystemPlugins\Wox.Plugin.SystemPlugins.csproj">
       <Project>{69ce0206-cb41-453d-88af-df86092ef9b8}</Project>
